windows-based pkgpanda builds
This commit enables builds of a small subset of packages on windows and
passes all the unittests that are currently ported to windows. Others
are disabled until the necessary parts of pkgpanda are ported.
Windows builds using a tree-variant of 'windows'.
Note: The tar files generated on windows have the binaries in them, but
at this time there is no configuration. Only placeholder configuration
exists. At this time package dependencies are not enabled on Windows.